Reorganized case-details folder structure and naming convention
