+++ web/src/components/case-details/civil/CivilSidePanel.vue
@@ -7,7 +7,7 @@
 <script setup lang="ts">
   import { civilFileDetailsType } from '@/types/civil/jsonTypes';
   import { AdjudicatorRestrictionsInfoType } from '@/types/common';
-  import AdjudicatorRestrictionsPanel from './AdjudicatorRestrictions/AdjudicatorRestrictionsPanel.vue';
+  import AdjudicatorRestrictionsPanel from '@/components/case-details/common/adjudicator-restrictions/AdjudicatorRestrictionsPanel.vue';
 
   const props = defineProps<{
     details: civilFileDetailsType;
