+++ web/src/components/case-details/common/AppearancesView.vue
