+++ web/src/components/case-details/common/CaseHeader.vue
