+++ web/src/components/case-details/common/CourtFilesSelector.vue
