+++ web/src/components/case-details/common/DivisionBadge.vue
