+++ web/src/components/case-details/common/DocumentsView.vue
