+++ web/src/components/case-details/common/Summary.vue
