+++ web/src/components/case-details/common/accused/Accused.vue
