+++ web/src/components/case-details/common/accused/AccusedPanel.vue
