+++ web/src/components/case-details/common/accused/Bans.vue
