+++ web/src/components/case-details/common/adjudicator-restrictions/AdjudicatorRestriction.vue
