+++ web/src/components/case-details/common/adjudicator-restrictions/AdjudicatorRestrictionsPanel.vue
